fix: make the order rounding rule configurable (#3801)
* fix: let a shop round the line total instead of the unit price

* test: cover both order rounding modes from cart to placed order

* fix: freeze the total of orders placed before a rounding mode switch

* test: cover the rounding pivot, discounts, postage and mixed tax rates

* fix: expose order unit amounts with the precision the order total uses

* fix: keep the three totals of an order line in agreement under rounding of sums

* fix: stop the api rounding order line unit amounts away and truncating a weight quantity

* feat: add a command that switches the order rounding mode and freezes past orders

* test: cover the line total identity, the api unit amounts and the switch command

* fix: stop the rounding mode command from moving the pivot when the mode is unchanged

+++ core/lib/Thelia/Api/Resource/CartItem.php
@@ -90,7 +90,9 @@ class CartItem implements PropelResourceInterface
 
     #[Groups([self::GROUP_ADMIN_READ, Cart::GROUP_ADMIN_READ, self::GROUP_FRONT_READ, Cart::GROUP_FRONT_READ, self::GROUP_FRONT_WRITE])]
     #[NotNull(groups: [Order::GROUP_ADMIN_WRITE])]
-    public ?int $quantity = null;
+    // cart_item.quantity is a FLOAT, like order_product.quantity: a line of
+    // goods sold by weight carries grams, not units.
+    public ?float $quantity = null;
 
     #[Relation(targetResource: Product::class)]
     #[Groups([self::GROUP_ADMIN_READ, self::GROUP_FRONT_READ])]
@@ -162,12 +164,12 @@ class CartItem implements PropelResourceInterface
         return $this;
     }
 
-    public function getQuantity(): ?int
+    public function getQuantity(): ?float
     {
         return $this->quantity;
     }
 
-    public function setQuantity(?int $quantity): self
+    public function setQuantity(?float $quantity): self
     {
         $this->quantity = $quantity;
